Fix Package search bugs
Change the set of package names that the package_autocomplete was
generated from to only select ones which have information in
PackageCategories.json.

Add a clearHighlighting function to be called both on the clear of the
autocomplete and the selection of a new package.  This prevents highlight
lines from being left behind if the search box didn't lose focus.