Add node for "New Requirement"
Add a purple colored node which will represent an entirely new Requirement
node to the BFF display to differentiate it from a node that was changed
recently.  Update the legend accordingly as well.

Add an additional filter to remove the first level of "children" if they
do not have any new or updated requirements