Add DOX and files URL data
Add a new file, created during the run of setup.py, which contains the
relative paths between the files directory and the DOX directory.

This **should** be the same relative path from the vivian html files to
each respective folder.

Start to transition each HTML page to use the new variables